requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.0
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==1.0.6
Tempita==0.5.2
Werkzeug==0.10.4
argparse==1.2.1
decorator==3.4.2
itsdangerous==0.24
pbr==1.3.0
six==1.9.0
sqlalchemy-migrate==0.9.7
sqlparse==0.1.15
wsgiref==0.1.2
